requirements.txt
fastapi
uvicorn
spleeter
tensorflow-cpu==2.9.0
ffmpeg-python
librosa
musdb
norbert
numpy
pandas
tqdm
typing-extensions>=4.12.0
wheel
responses
python-multipart

